+++ v3/crates/compatibility/src/compatibility_date.rs
@@ -159,7 +159,8 @@ pub fn get_compatibility_date_for_flag(flag: Flag) -> Option<CompatibilityDate> 
         | Flag::DisallowDataConnectorScalarTypesMismatch
         | Flag::CheckObjectTypeFieldsExist
         | Flag::DisallowOrderByFieldsWithFieldArguments => Some(new_compatibility_date(2025, 2, 4)),
-        Flag::DisallowUnsupportedOrderableRelationships => {
+        Flag::DisallowUnsupportedOrderableRelationships
+        | Flag::DisallowLocalRelationshipsOnDataConnectorsWithoutRelationshipsOrVariables => {
             // TODO Set a compatibility date during the next release
             None
         }